pyproject.toml
[project]
name = "koubou-system"
version = "0.1.0"
description = "工房システム POC - AIワーカープール管理システム"
requires-python = ">=3.9"
dependencies = [
    # Web Framework & API
    "flask>=3.0.0",
    "flask-cors>=6.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "ariadne>=0.26.0",  # GraphQL
    
    # HTTP & WebSocket
    "requests>=2.31.0",
    "httpx>=0.28.0",
    "websockets>=15.0.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    
    # Async Support
    "anyio>=3.7.0",
    "pytest-asyncio>=1.1.0",
    
    # Configuration & Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "click>=8.0.0",
    
    # Data Validation
    "pydantic>=2.0.0",
    "annotated-types>=0.7.0",
    
    # LLM Integration
    "ollama>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.1.0",
    "black>=25.0.0",
    "flake8>=7.3.0",
    "mypy>=1.5.0",
]

test = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.28.0",  # For testing HTTP endpoints
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [".koubou"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.1.0",
    "black>=25.0.0",
    "flake8>=7.3.0",
    "mypy>=1.5.0",
]
